Epic5-106: Переместила CountryRegionData в domain -> models
