+++ app/src/main/java/ru/practicum/android/diploma/domain/models/CountryRegionData.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.domain.models
+
+data class CountryRegionData(
+    val countryId: String,
+    val countryName: String,
+    val regionId: String,
+    val regionName: String
+)
