+++ app/src/main/java/ru/practicum/android/diploma/util/diffutils/CountryRegionDiffCallback.kt
@@ -1,7 +1,7 @@
 package ru.practicum.android.diploma.util.diffutils
 
 import androidx.recyclerview.widget.DiffUtil
-import ru.practicum.android.diploma.ui.filter.region.viewmodel.CountryRegionData
+import ru.practicum.android.diploma.domain.models.CountryRegionData
 
 class CountryRegionDiffCallback : DiffUtil.ItemCallback<CountryRegionData>() {
     override fun areItemsTheSame(oldItem: CountryRegionData, newItem: CountryRegionData): Boolean {
